Shim swanky-generic-array to point at the generic-array types
Before this upgrade of the generic-array crate, ArrayLength was
parametric over the type contained in the array. We wrote wrappers for
Swanky which used GATs to allow generic-arrays to hold any type.

The latest version of generic-array includes this feature, so this
commit turns swanky-generic array into an alias to the new types.