Update uses of generic-array for the new version
Aside from changing API names, and the removal of the type argument to
the ArrayLength trait, the major change required when upgrading
generic-array is that the rustcrypto ecosystem hasn't upgraded yet. As a
result, when interacting with that ecosystem, we need to use _their_
version of generic_array (which they helpfully re-export).