svole_trait.rs: Update definition of SvoleT trait.
NOTE: This commit does not build! This is a very small chunk of a
big refactor that touched most of Diet Mac'n'Cheese, committed as
such to make code review easier. Start here for everything to make
sense!

Previously, the SvoleT trait was defined over a single generic
parameter, M. This, in conjunction with not being "party-aware",
meant that every other structure in Diet Mac'n'Cheese that used
SVOLE had a significant amount of redundancy in the generic
parameters (to reconcile the need for a prover-SVOLE and a
verifier-SVOLE).

This commit changes the definition of SvoleT in a few ways:

- It is now generic over a Party, the value field, and the tag
  field
- Because Diet Mac'n'Cheese uses ocelot's SVOLE implementation
  under the hood, the extend method is changed to explicitly
  support the two types of vectors that ocelot's API deals with.
  While a bit clunky, this does maintain the previous version's
  property of optimizing memory usage by not relying on
  intermediate vectors in implementations of the method.
- The delta method is made verifier-only through an evidence
  parameter

These changes will reduce the redundancies introduced in earlier
work, and enable further redundancy reduction in the definitions
of structures and their associated algorithms.