homcom.rs: Use the updated SvoleT for homomorphic commitments.
NOTE: This commit will not build! This is just a small chunk of a
large set of refactors applied to Diet Mac'n'Cheese intended to
make code review easier. These messages should be read in order!

The first point of redundancy that can be removed using the party-
aware version of the SvoleT trait is in the context of homomorphic
commitments. Where we previously needed two SVOLE generics, we now
only need one (since implementors of SvoleT must also be party-
aware).