homcom.rs: Use VerifierPrivate for the open method.
This commit updates the implementation of the FCom structure to
use a VerifierPrivate Vec in the open method, since provers do
not need this output vector.

Note that this commit does _not_ propagate this change to call
sites, as there are more substantial refactors to commit that
make those specific changes hard to isolate.