svole_trait.rs: Combine the SvoleSender and SvoleReceiver types.
This commit constitutes the remaining changes to the svole_trait
module, in particular de-duplicating the SvoleSender/SvoleReceiver
types using the now party-aware SvoleT trait.

This is analogous to changes in the prior MR to the FCom and Mac
types, and is the change necessary to, in concrete contexts,
properly remove the previously-introduced redundancy in SVOLE.

This commit does _not_ propagate this change in type, and will
thus not build in isolation.