svole_thread.rs: Make SvoleAtomic party-aware.
This commit changes SvoleAtomic to take the same generic parameters
as the SvoleT trait, and updates the fields of the structure to
match.

One surprise may be that delta does not change type here. This is
due to the fact that the 'polling' loop in the implementation of
the SvoleT delta method is crucial to keeping the threads
synchronized - something like Option is necessary to loop until the
value has been set by this or another thread.

Otherwise this commit contains a few minor adjustments to
documentation.