edabits.rs: De-duplicate with the Party abstraction.
This is a very large commit, collapsing all of the redundant
structures used to implement the edabits protocol.

This code will be worth revisiting for further simplification in
the future, but in summary, all structures and algorithms have been
'unified' in the same was as Mac, FCom, etc.

This does result in an overall reduction in the SLOC, and reduces
the potential for 'drift' in the two parties' implementations
(since of course there is now only one implementation that is
party-specific only where it needs to be).

The changes to this implementation have been gone over with a fine-
toothed comb by both Benoit and myself (since there was some bug
causing testing failures recently that we have since fixed). As
it will be a target of future refactoring (as one of the more
complicated parts of DMC), we defer detailed explanations of the
changes.