backend_trait.rs: Remove the Party enum and use swanky-party.
This commit makes the enum Party (and corresponding method on the
BackendT trait) obsolete, instead parameterizing the trait on the
party.

In the future we may be able to make more robust use of this to
make a more idiomatic API for BackendT (e.g. making input_private
require evidence that the party is Prover), but this refactor only
requires that the party be accessible to anything making use of the
BackendT trait.