backend.rs: Unify into a single DietMacAndCheese structure.
This commit is one of the big wins of this refactor: Rather than
defining separate Prover and Verifier versions of the
DietMacAndCheese protocol, we now have a single structure (and
associated implementations) that unifies the behaviors of provers
and verifiers, using the abstraction to localize the places where
behavior differs.

After this, there is very little that has not been made generic
over the running party. This will be used at the topmost level to
determine all downstream behaviors as appropriate.