private.rs: Add mapping method that provides evidence to ProverPrivate.
This small change to the swanky-party crate adds a version of map
that takes the evidence of the party in the function. This is
convenient for some situations, where the body of the function
being mapped calls functions that are prover-only (i.e. require
evidence that the party is Prover).