private.rs: Remove the map_with_ev method.
We previously extended the swanky-party ProverPrivate types with
a mapping method that provides evidence of being the prover to the
function being mapped.

This seemed convenient (as we may want to call prover-only APIs in
the mapped function's body), but is somewhat awkward.

An alternative solution, more appropriate in a language like Rust
(without type families / type-level constants), is to change the
definition of the value() method on MACs to return a ProverPrivate
value, rather than taking evidence directly.

This commit implements these changes, which allows map_with_ev to
be removed (thus preventing us from changing multiple crates in
this one MR).