backend_trait.rs: Remove the generic Party parameter.
When refactoring DMC to use swanky-party, the BackendT trait was
modified to carry a generic parameter bounded by Party.

This was not actually necessary; while the party must be known for
some plugins to operate correctly (and implementations of the trait
are party-aware), this can be handled in a more localized way (that
is, via generic parameters on the plugin execution functions).

This commit removes P: Party from the definition of BackendT, and
updates all effected types and functions to match. This has
dramatically simplified a good amount of the Dora code, and there
is generally less 'clutter' in the form of type instantiations or
specializations of functions to particular Party types.

Given the nature of the change, it is captured in a single commit.