mac.rs: Add blanket implementation of MacT for FiniteFields.
To demonstrate that MacT does not actually limit the use of
BackendT in a meaningful way, this commit adds a blanket
implementation of the MacT trait for all T: FiniteField.

In particular, it suffices to set the Value and Tag associated
types to the type itself. Furthermore, the LiftedMac is just the
type itself, as there is nothing _to_ lift.

Finally, the implementation of lift() asserts that the given array
has length 1 (it must, since DegreeModulo<T, T> == 1), and just
returns this one element.