Use tree-sitter to check for deny(missing_docs)
Previously, we checked for #![deny(missing_docs)] using a regex, and
said that the crate was good if it contained (some variation of) that
string.

During development, I sometimes comment out the deny(missing_docs)
directive, and I want the lint to remind me to un-comment it out before
pushing.

To do this, we use the tree-sitter parsing framework to query for the
deny(missing_docs) attribute, and check that. Tree-sitter is smart
enough to properly parse out comments, and so will not be confused by my
dastardly comments!