Construct TextRelation from a type store AND function store.
Previously, a new TextRelation was created given only a type
store.

To support the correct propagation of function declarations to
functions defined in the circuit, this commit changes the new
associated method to take a function store in addition to the
type store.

This was a bug causing ciruits with nested function calls to fail,
found by running the University of Vermont IRS circuit.