scuttlebutt: Remove AbstractChannel::clone
AbstractChannel::clone is used in very few places in swanky.

It shouldn't be used at all.

Properly using a channel requires exclusive access by the user. This is
best modeled by an &mut reference which guarantees unique access to the
channel. AbstractChannel::clone should not be used for the same reason
that &mut references can't be clone in general--it breaks the uniqueness
guarantee.

Beyond that, the mechanisms that we use to clone abstract channels
(Rc/Arc and RefCell/Mutex) incur substantial overhead on each
_high-level_ write operation (i.e. the cost is not amortized across the
size of read and write buffers).