ram_v1.rs: Initial frontend structure for the RAM plugin.
This commit implements the scaffolding to support the SIEVE IR
RAM plugin in Diet Mac'n'Cheese. This includes the addition of
two Plugin types for each variant of the plugin (ram_bool_v1 and
ram_arith_v1), and a wrapper type encapsulating both variants
(since actual implementation of the Dora RAM protocol is agnostic
to the underlying field).