lib.rs: Add as_single_wire method to WireRange.
This commit adds the method as_single_wire to WireRange, which
checks that the WireRange is in fact a single wire and returns the
result if so.

This simplified the check_wire_range_buf_single_output method,
and led to the public/private input methods being correct (though
at the cost of adding a call to as_single_wire everywhere the other
method was originally being used).