circuit_ir.rs: Update gate representation for latest circuit IR.
NOTE: Does not build!

This commit updates the definition of GateM for the latest SIEVE
circuit IR - namely, to provide multi-wire copy gates and public/
private inputs.

This included updates to the DMC-specific text and flatbuffer
parsing code, and updates to some methods that previously assumed a
single wire for those gates.

Note that this does _not_ provide updates to the backend
implementations or plugins; these changes will come in future
commits.