backend_multifield: Implement multi-wire copy functionality.
This commit implements multi-wire copy in the 'obvious' way: Use
the already-implemented single-wire copy on all of the input wires.

This avoids changing the definitions of the trait BackendT or any
of its implementations.

We should consider changing this in the future to reap the benefits
of multi-wire copying (ie, doing some careful memcpying that takes
advantage of our wire memory management).

That said, there should be no additional cost on any circuits we
currently test with, as these were already using the older copy
gate that only operates on single wires.