backend_multified.rs: Implement multi-public/private inputs.
The updated SIEVE IR allows the public/private input gates to
output to a range of wires. This commit implements that
functionality.

Similarly to the multi-wire copy, this implementation merely makes
repeated calls to the backend public/private input method. In this
case, there probably isn't much to do better unless we get a bit
fancier with our queues of input values (eg, to allow more
efficient popping of many values at once).