dora/mod.rs: Update the translation to disjunction gates.
Dora uses an internal representation of gates that differs from
that of Diet Mac'n'Cheese. In particular, Copy and Witness gates
only support single wires in the Dora context.

Rather than changing the type definition (and subsequently a _lot_
of other code), this commit simply updates the translation code to
emit a series of Dora gates for Copy and Witness DMC gates. To
avoid allocation, this involved moving the definition of translate,
passing a mutable reference to the output, and rearranging a few
other small helper functions.