backend_multified.rs: Fix bug in witness gate implementation.
Thanks to Benoit for catching this very subtle bug!

Previously, we looped over the (possibly None) collection of
witness values to implement private inputs, to mirror the behavior
of public inputs.

This is incorrect, as in the case that the witness collection is
in fact None (ie, the verifier is executing), the verifier still
needs to execute the backend's input_private method and set the
memory. This caused many test failures due to uninitialized wire
memory being accessed.

This commit instead loops over the output wires, conditionally
indexing into the collection of witness values.