backend_multifield.rs: Implement initial evaluation of RAM ops.
RAM operations are going to be evaluated by the evaluator
associated with the underlying address/value field of the RAM being
operated on / initialized. This is because we must access the wire
memory associated with that field simultaneous with the state
associated with RAM operations.

This commit is a small step towards that deferred evaluation,
removing the todo! for the RAM PluginExecution type in the
evaluate_call_gate method.