backend_multifield.rs: Add scaffolding for evaluation of RAM ops.
This commit simply adds the pattern-match scaffolding for RAM
plugin operations to the plugin_call_gate method.

For now this assumes that both Boolean and arithmetic RAMs will be
handled by the same code; this may change in future commits since
Boolean RAMS support multi-wire addresses/values, but there is no
current indication that this will be necessary.