backend_multified.rs: Add sanity checks and updated TODOs for RAM.
This commit adds sanity checks to plugin_call_gate for all RAM
operations - in particular checking the number and width of wire
ranges that are inputs/outputs to the operations.

Additionally, the TODO messages have been updated to be more
specific to what remains for these operations' implementations.
This is an improvement on previous commits which used generic
language about implementing the features, rather than anything
specific about _how_ they are going to be implemented.