backend_multifield.rs: Implement a 'dummy' evaluator.
Diet Mac'n'Cheese implements a form of dynamic dispatch for circuit
evaluation: For each type used in a circuit, a backend implementing
EvaluatorT is loaded to handle evaluation of gates over that type.

There are some types, however, such as RAM, that do _not_ need to
implement their own evaluation, as this is handled by/with the
backend associated with the RAM's address/value type.

Despite this, an evaluator _must_ be provided for such types, as
the initialization of DMC currently halts for any non-field type.

This commit provides a new evalaution type, EvaluatorDummy, that
can be used for such types. This type is a unit-like struct whose
implementation of EvaluatorT simply panics on all operations (as
again, this is never intended to actually be invoked during circuit
evaluation).