backend_multifield.rs: Load dummy evaluators for RAM types.
(See the commit log for details on the need for and implementation
of the EvaluatorDummy.)

This commit adds validation of RAM types in the backend-loading
code, and for those types pushes an instance of EvaluatorDummy.

The validation means Diet Mac'n'Cheese will abort early on
circuits containing ill-formed RAM types, and the addition of
dummy evaluators will make sure that there is consistency between
the type IDs implicitly assigned to types declared in the circuit
and the evaluators associated with them - even if these evaluators,
as mentioned in a previous commit, will never actually be invoked.
This consistency is crucial to the correctness of DMC.