src/ram: Add directory for Dora RAM implementation.
This commit adds a new directory for the implementation of the
Dora RAM protocol, including an adapted implementation of the
transmission channel (for Fiat-Shamir).

This is the beginning of the most important part of implementing
the SIEVE RAM plugin: Adapting Mathias' Dora implementation to use
the Party abstraction and be more 'dynamic' to support the needs of
the SIEVE IR.

As a start, this includes the transmission channel, which now uses
the updated definition of AbstractChannel & returns challenges more
dynamically (where previously const generics were used for
challenge value generation, we now use dynamic/runtime parameters
and heap-allocated vectors. This is due to the 'dynamic' nature of
the SIEVE plugin, where there is more freedom than originally
assumed when creating RAM types/values in circuits.