protocol.rs: Initial struct definition for RAM protocol.
This commit provides the first significant departure from Mathias'
implementation of RAM. Here, we begin to unify the separate prover/
verifier structures using Party, and remove the use of const
generics in favor of more Vecs/dynamic structure.

The former simplifies the code in the same way disjunctions (and
the rest of DMC) were previously simplified: Reducing duplication
between prover/verifier, and making it clear where party-specific
behavior is being implemented.

The latter is a result of the SIEVE RAM plugin being significantly
more dynamic than was assumed for the initial implementation. In
particular, we had to abandon the use of const generics, since we
only learn the address/value/challenge sizes at runtime. To enable
this, we replace the fixed-size arrays with Vecs, and remove all
const generics.

These const generics may need to be added back as fields of the
structure, depending on the needs in the actual protocol
implementation; this commit does not assume that need is present.