protocol.rs: Implement potential bugfix for CI/RAM test failure.
The test for RAM operations was failing due to an out-of-bounds
panic on the flattened vectors at the heart of the protocol's main
loop. This was due to attempting to access the vectors by index,
rather than pushing elements to them directly.

This commit updates the loops to push to the vectors, with the
added bonus of removing a now-extraneous call to enumerate().