protocol.rs: Fix bug in initial allocation of flattened Vec.
Previously, we were encountering a commitment error. This was
determined to be due to the flattened vector (containing the
address, value, and challenge values) being too large.

The reason the vector was too large was its initialization: Rather
than simply pre-allocating enough space for the flattened vector,
we allocated _and_ initialized, meaning it contained extra values
before the loop adding values was run.

This commit corrects that error by replacing the macro call vec!
with Vec::with_capacity, mirroring the other methods on DoraRam.