ram_v1.rs: Use usize for the RAM size given to init.
Previously, we passed along the Number (Uint<...>) for the size of
the RAM to initialize.

This isn't actually useful, since we can't feasibly support RAMs
with a number of cells larger than u64's maximum value. This commit
reflects this reality by saving the RAM size as a usize, which
matches the implementation of the Dora protocol.