ram/mod.rs: Add stub BinaryRam type and Ram wrapper-type.
Eventually, to support RAMs over binary/F2 addresses and values,
we will need a structure analogous to ArithmeticRam.

This commit adds a unit-like struct for this to-be-defined binary
RAM type, as well as a wrapper type, Ram, that encapsulates both.
This will enable RAM states to be stored in the same way "higher
up" in Diet Mac'n'Cheese.