backend_multifield.rs: Add ram_states to DietMacAndCheeseConv.
To support RAMs over arithmetic fields, we must store the protocol
states for the RAMs somewhere. For arithmetic RAMs, the most
natural choice is DietMacAndCheeseConv, which is the BackendT used
for these fields.

The ram_states Vec should be thought of as a _store_, where the
indices act as 'pointer' values that SIEVE wires of RAM type carry.
This store is implicitly globally-scoped, meaning all RAM-typed
wires (regardless of their scope as defined by the SIEVE IR) will
refer to these global RAMs (which is consistent with the RAM plugin
specification).