ram/mod.rs: Remove the Ram wrapper type.
Originally, it seemed as if it was appropriate to account for all RAMs
in a single type. This was not the case, however, considering how
Diet Mac'n'Cheese backends are loaded: Arithmetic backends are loaded
as DietMacAndCheeseConv, while F2 is loaded as
DietMacAndCheeseExtField.

This commit removes that type, and acts as something of an extension
to the previous commit which used ArithmeticRam for the RAM states.