backend_multifield.rs: Improve BackendRamT interface.
First, this commit introduces the type alias RamId. This alias is used
to represent the unique identifiers RAMs will have (per underlying
field type), ie the values that RAM-typed wires in circuits will
carry.

Next, this commit updates the documentation for and signatures of the
methods BackendRamT types must implement to include an argument for
the RamId to act on (and, in the case of the init_ram method, a RamId
return value and init_value argument to match the SIEVE
specification). There's also now consistent use of the type
eyre::Result rather than the Rust std Result.