backend_multifield.rs: Require RAM support in EvaluatorSingle.
To execute RAM operations, we require that the backend type used in an
EvaluatorSingle implement BackendRamT, mirroring the other plugin and
feature backends (eg disjunctions & field conversions).

To allow Diet Mac'n'Cheese to compile with this change, we
additionally add stub implementations of BackendRamT for the types
DietMacAndCheeseConv, DietMacAndCheeseExtField, and
DietMacAndCheesePlaintext. The first two of these will be replaced
with actual implementations; they panic with todo! to reflect
this. The last (the plaintext evaluator), as of this writing, has no
plan in place to support plaintext RAM operations (though this would
be relatively easy to implement).