backend_multifield.rs: Add a Memory to map wires to RAM IDs.
Recall that RAM-typed wires do not have their own evaluation strategy,
instead being provided instances of EvaluatorDummy that will never be
invoked.

Operations on a RAM are delegated to the evaluator for the underlying
field type - that is, the type of the addresses/values stored in the
RAM. But, we still need to contend with the wires of RAM type. In
particular, they are to act as references to RAMs created with a
plugin-defined init function.

This commit adds a field ram_wires to EvaluatorSingle (the evaluation
strategy for a specific field), which is a Memory mapping these wires
to RamIds. Through this structure, we can appropriately track which
globally-scoped RAMs are being referenced by wires..