backend_multifield: Remove panics from some EvaluatorDummy methods.
The callframe_start and callframe_end methods on the circuit evaluator
are defined to use the evaluator for each type present in a function
call to push and pop frames, which challenges an earlier idea that
EvaluatorDummy would _never_ be invoked.

This commit updates the push_frame and pop_frame methods in the
implementation of EvaluatorT for EvaluatorDummy to do nothing rather
than panic, and documents this property. The other methods are left as
panicking/unimplemented!, as they should truly never be invoked.