backend_multifield: Implement the remaining todo! items for RAM.
This commit is the final "hooking together" necessary for RAM to be
properly interpreted in real circuits.

Here, we set our ram_wires in the handling of init. This should
correctly account for function scoped through the use of push_frame
and pop_frame on ram_wires, which happens simultaneously with the
pushing/popping of frames for the address/value field wires.

Note that the delegation to the backend is key: This is how this code
can work for F2 and arithmetic fields, as the associated structures
will delegate appropriately to ArithmeticRam or BinaryRam (which has
not yet been implemented).