backend_multifield: Change finalize_ram to finalize_rams.
Previously, we exposed finalize_ram in BackendRamT to take the unique
identifier for the RAM to finalize.

Since finalization should happen all at once for any RAMs, this commit
changes this interface to expose the method finalize_rams, which has
been implemented for DietMacAndCheeseConv in the obvious way (that is:
iterate over the RAMs, finalizing each of them).