ram_v1: Store the RAM type ID in RamV1.
In an effort to resolve some problems with the current implementation
of RAM (particularly when it comes to appropriately managing the
wires), this commit adds the field (and associated getter on
RamVersion) ram_type_id to RamV1, which holds the SIEVE type
identifier for the RAM type being operated on.

We need access to this type ID in evaluate_call_gate, which was the
primary motivation for this change.