backend_multifield: Replace EvaluatorDummy with EvaluatorRam.
Previously, we used a 'dummy' EvaluatorT type to handle any evaluation
over RAM-typed wires. First, this was insufficient: That
implementation panicked in methods that will be called for _all_ types
implementing EvaluatorT, such as those related to wire memory and
callframes.

Second, the need for a RAM-specific evaluator has become evident due
to the use of RAM-typed wires. It is not sufficient to manage these
wires from within the underlying field's evaluator, particularly when
it comes to allocate_slice / allocate_new.

These changes do imply some significant restructuring of the code,
particularly when it comes to the acquisition and use of RamIds. These
changes will be necessitated by the removal of ram_wires from
EvaluatorSingle, which will complement this commit.