backend_multifield: Update the signature of plugin_call_gate.
While RAM-typed evaluators are responsible for managing the wire
memory associated with RAMs, the evaluator for the underlying
address/value field is responsible for actually calling out to the RAM
protocol (since, of course, the protocol only cares about values
of/composed of values of this underlying field type).

This poses something of a problem: We must share RamId information
between evaluators that can only interact via the circuit-level
evaluator that owns them.

The underlying address/value field evaluator is a producer of RamIds,
via calls to the backend's init_ram method that RAM initialization
calls to allocate RAM protocol state. But it is also a consumer, since
the ram_read and ram_write operations require RamIds.

Since these RamIds are produced and consumed by the plugin methods
associated with the address/value field, this commit changes
plugin_call_gate to optionally take/return a RamId.

As noted in the documentation, this is probably not the best long-term
design; an enum PluginExtras that is Option-like would be more
appropriate, in anticipation of other plugins needing to take or
return auxiliary information.

This commit leaves a todo! to update the implementation of
evaluate_call_gate to properly work with RAM-typed wires and RamIds.