backend_multifield: Address todo! in RAM operation evaluation.
Previously, evaluate_call_gate naively delegated to the underlying
address/value field backend for all RAM operations. This left a todo!
in the call to plugin_call_gate, to provide the appropriate RamId for
reads/writes (and account for RamIds returned by inits).

This commit replaces this naive delegation with operation-specific
behavior that brings together the EvaluatorRam and EvaluatorSingle
associated with these operations. The handling of init now delegates
to the underlying field, saving this in the proper RAM wire memory,
using its EvaluatorRam. Read and write analogously fetch the RamId
using the EvaluatorRam, then pass this to plugin_call_gate to
ultimately delegate evaluation responsibilty to the appropriate
field's EvaluatorSingle.

Something really nice about this change is that we no longer need to
consider anything strange to account for the possibility of multilple
binary-backed RAM types; RAM wire memory management now takes
advantage of type-scoping, so we get that handling for free.