backend_multifield: Fix the implementation of callframe_start.
The SIEVE IR specification says of functions:

> When the function is invoked, each range is mapped into its scope
  sequentially from 0.

That is: On a per-type basis, output (then input) wires to a function
are numbered starting from 0.

The previous implementation of callframe_start did not number function
wires in this way; rather, it treated every wire the same, and
numbered them all incremenetally from 0, regardless of type.

This commit corrects this error, instead maintaining a map from types
to current wire total for use in allocation of slices. This change
results in a simple RAM test circuit (involving RAM operations inside
of function calls) successfully running to completion.