diet_mac_and_cheese: Consistently important eyre types/macros.
Previously, there was a mix of qualifying and importing the types and
macros from eyre. This commit updates all of Diet Mac'n'Cheese to
consistently import these types and macros rather than using full
qualification (with the exception of test code where eyre is used;
here, full qualification is still used).