ram/protocol: Support specification of initial RAM values.
The previous implementation initialized all RAMs with all-zero
values. The SIEVE RAM specification, however, allows for the initial
value of a newly-allocated RAM's cells to be specified.

This commit adds support for this initial value in general for RAMs
with any address/value size, and specifically for ArithmeticRam.

In short, initial values need to specified in two places: In the
prover-private actual representation of memory (ie the map from
addresses to values/challenges), and in the RAM finalization's
construction of 'pre' values used to set up the permutation check.

To minimize allocations, we keep the definitions of pre similar to
what they were before, adding in a copy_from_slice to splice in the
initial value (that we augment with default challenge values upon RAM
initialization).