ram/protocol: Improve the protocol documentation.
This commit adds a (very) high-level description of the RAM protocol,
in terms of how the methods on DoraRam are used to implement reads and
writes (and how the commitments associated with these operations are
validated).

It also links to the paper the implementation is based on, and
provides some informative sentences to remove/insert about their
intended use.