fancy-examples/sum: Return and open the result of the fancy garbling computation
Summary: The Garbler and the Evaluator need to jointly "reveal" the output of
the garbled circuit since it would otherwise be "garbled" (i.e. in some sense
encrypted).

Additionally, the Evaluator receives the output of the computation and
transforms it from a base 2 bit vector to base 10 u128 for ease of readability.