fancy-examples/sum: Test the fancy circuit result against the in the clear result
Summary: The Garbler and the Evaluator are turned into seperate threads with
their own channels and rng, that each seperately call their own fancy method.
The result of the fancy circuit is then compared against the insecure/in clear
result that we expect.

We need to split the Garbler and the Evaluator into seperate threads to emulate
how in practice each party would be operating on a seperate machine with a
seperate state. We also need to validate that the fancy circuit is outputting
the correct result by comparing it with the output of the "insecure" equivalent
summation function.