fancy-examples/sum: Use clap to parse the example inputs
Summary: Use the clap parser to parse the example command line inputs.

Without clap, we would need to manually parse the command line input from a
String to whatever our fancy circuit expects the input to be. The fancy
examples each expect a different type of input and clap alleviates the burden
of handling each seperately.