fancy-examples/gcd: Creates a simple in the clear gcd example
Summary: Defines the ideal functionality that we will implement in fancy garbling.
Takes two numbers from the garbler and the evaluator passed as command line arguments, computes their gcd and returns the result.

We use the "in the clear" computation to verify that the output of fancy is done
correctly and define the ideal behavior that we would like the garbled circuit to implement.