fancy-examples/linear_oram: Creates a simple in the clear linear oram example
Summary: Defines the ideal functionality that we will implement in fancy garbling.
Takes an index and array, finds the value of the array at that index and return
the result.

We use the "in the clear" computation to verify that the output of fancy is done
correctly and define the ideal behavior that we would like the garbled circuit to implement.