fancy-examples/linear_oram: Creates Garbler and Evaluator
Summary: Creates the Garbler and Evaluator via their own channel and rng.

The fancy function on its own does not define the behavior of the Garbler and
the Evaluator prior to the fancy function call. We hence need to create and
define what the Garbler and Evaluator do prior to that call.