circuit_ir: Add a SieveIrDeserialize trait.
Currently, much of Diet Mac'n'Cheese is overly-constrained to
PrimeFiniteFields, due to the fact that only PrimeFiniteFields provide
the tools necessary to convert to/from big integer representations.

SIEVE IR supports integer representations of extension field values,
and we must support these representations for the implementation of
extension field type support and the lifting required for the security
of some operations.

This commit takes the first step toward addressing this need by
implementing a trait, SieveIrDeserialize, that will encapsulate the
ability to deserialize from a Number (the representation of constant
values in SIEVE IR).