circuit_ir: Implement SieveIrDeserialize for all PrimeFiniteField.
As we are aiming to relax the constraints in Diet Mac'n'Cheese to
allow for the use of extension fields, the commit prepares for that
constraint-relaxation by implementing the new deserialization trait
for all PrimeFiniteFields.

Note that this implementation is identical to that on PrimeBackendT, a
trait that will eventually be removed/replaced with the functionality
provided by SieveIrDeserialize.