dora: Remove PrimeFiniteField constraint in Dora.
Because Dora must compile to R1CS (and in the process deserialize
SIEVE IR constants into FiniteField values), it was constrained to
operation over PrimeFiniteField, to get access to the try_from_int
method.

This commit replaces that constraint with FiniteField +
SieveIrDeserialize, which allows Dora to operate over more types -
notably, any extension field that we implement deserialization for.

This will be important for the implementation of binary disjunctions.