circuit_ir: Remove blanket SieveIrDeserialize for PrimeFiniteField.
In a previous MR (!388), the SieveIrDeserialize trait was added to
encompass copy-able types for which values could be deserialized from
SIEVE IR constants. A blanket implementation for PrimeFiniteField
types was provided, so that the existing code would be as undisturbed
as possible by the change.

Unfortunately, limitations of Rust meant this blanket implementation
could not be left in place. In particular, Rust complained when adding
an impl SieveIrDeserialize for F40b, since it is possible that future
versions of the swanky-field-binary crate may add a PrimeFiniteField
instance for F40b - this possibility is enough to trigger an
"overlapping instances" error.

Now, we _know_ that such an instance will never exist (since, of
course, F40b is not a prime field); but there is currently no way
in (stable) Rust to express these kinds of 'negative' trait
bounds (see the unstable Rust book on negative impls and
specialization).

So, to remove the blanket implementation, this commit implements a
macro impl_sieve_ir_deserialize_prime_field that handles the
boilerplate implementation of the deserialization for types that are
PrimeFiniteField, and calls this macro for all prime fields currently
supported by Diet Mac'n'Cheese.

This is of course not _completely_ ideal; now, we must be sure to keep
the macro call up-to-date with fields.rs / the few other places where
manual synchronization is necessary. Furthermore, we needed to add
additional + SieveIrDeserialize bounds where previously
PrimeFiniteField would suffice; this is arguably better, though, since
it is more explicit about the deserialization requirement everywhere
it is necessary.

This commit also provides a stub todo! implementation of
SieveIrDeserialize for F40b, as a way to test the macro-based solution.