fields.rs: Move the SieveIrDeserialize trait.
Previously, the SieveIrDeserialize trait was defined (and implemented
for relevant types) in the circuit_ir module.

To keep the set of modules that must be changed when new fields are
supported small, this commit moves this trait and its implementations
to the fields module.

In addition, this commit replaces the stub implementation for F40b
with another macro that can be used to implement SieveIrDeserialize
for any binary extension field, and uses this macro to provide an
implementation for F40b.