dora/disjunction: Pass num_cond to Disjunction::compile.
In the Boolean case, we modify the number of condition wires when we
lift to the secure 40-bit field F40b from the number of F2 inputs to
1. This requires that we be able to specify the number of condition
wires when compiling a disjunction, as the original DisjunctionBody
will report the number of F2 wires.

This commit modifies the compile method to take num_cond, and updates
the arithmetic/prime field implementation to use the original
disj.cond() (which must be 1).